mvp/test/requirements.txt
kafka_python==2.0.1
Flask==1.1.2